Relocate existing tests to `MapOf`.
