+++ tests/Unit/MapOf/KeyBy/Alias.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit\MapVia;
+namespace Tests\Unit\MapOf\KeyBy;
 
 use Zerotoprod\DataModel\DataModel;
 
@@ -8,5 +8,6 @@ class Alias
 {
     use DataModel;
 
+    public string $id;
     public string $name;
 }
