+++ tests/Unit/MapOf/KeyBy/User.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit\Array;
+namespace Tests\Unit\MapOf\KeyBy;
 
 use Zerotoprod\DataModel\DataModel;
 use Zerotoprod\DataModel\Describe;
@@ -15,6 +15,7 @@ class User
     #[Describe([
         'cast' => [self::class, 'mapOf'],
         'type' => Alias::class,
+        'key_by' => 'id'
     ])]
     public array $Aliases;
 
@@ -23,6 +24,7 @@ class User
         'cast' => [self::class, 'mapOf'],
         'type' => Alias::class,
         'level' => 2,
+        'key_by' => 'id'
     ])]
     public array $AliasesNested;
 }
